Holiday Data

Builds a dataframe of holiday dates between 2 dates in the form of 'yyyy-mm-dd'. Pass a dataframe with the column of dates.

Usage

holiday_data(df, out_df = c("weeks", "days"))

Arguments

df

dataframe A dataframe with 1 date column with at least 2 dates.

out_df

character weeks is the default, but can also use days.

Value

The result of the function can be either a dataframe with days between the minimum and maximum of the dates provided, or, the default option which is to return a dataframe with weeks instead of days. The paramater out_df gives the user 2 choices for output. The default output a weeks dataframe with beginning week, ending week, a holiday flag, and a column for the holiday name. The second option returns a dataframe with all the days and day names with the 4 columns mention above.

Note

https://github.com/cran/timeDate/blob/master/R/holiday-Dates.R

https://www.federalpay.org/holidays

Examples

if (FALSE) {
weeks_df <- tibble::tibble(weeks = seq.Date(as.Date('2015-01-01'),
as.Date('2020-01-01'), by = 'week'))
holidays_weeks(df = weeks_df) %>% head(10)

## Individual Holidays
years <- seq(2015,2020, by = 1)
US_ValentinesDay <- USValentinesDay(years) %>% as.Date()
Name_Valentines <- Holiday_Names("ValentinesDay",US_ValentinesDay)
}